requirements.txt
selenium==4.41.0
pytest==9.0.2
allure-pytest==2.15.3
webdriver-manager==4.0.2
pytest-xdist==3.8.0
